refact: extract styles from components
