+++ src/components/Footer/styles.tsx
@@ -1,0 +1,13 @@
+import { Link } from "gatsby"
+import styled from "styled-components"
+
+export const FooterWrapper = styled.footer`
+  margin-top: 32px;
+  padding: 32px 0;
+  background: #171f26;
+`
+
+export const StyledLink = styled(Link)`
+  color: #9cb2ff;
+  text-decoration: none;
+`
